% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{aggregation}
\alias{aggregation}
\title{Defines a reference to a Druid aggregation}
\usage{
aggregation(x)
}
\arguments{
\item{x}{name of the aggregation}
}
\description{
Defines a reference to a Druid aggregation
}

