% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{cardinality}
\alias{cardinality}
\title{Constructs a cardinality aggregator for the given Druid columns}
\usage{
cardinality(name, fieldNames, byRow = F)
}
\arguments{
\item{fieldNames}{list of Druid columns}

\item{byRow}{setting byRow to false (the default) it computes the cardinality of the set composed of the union of all dimension values for all the given dimensions}
}
\description{
Constructs a cardinality aggregator for the given Druid columns
}

