% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dimension}
\alias{dimension}
\title{Creates a Druid dimension object}
\usage{
dimension(name)
}
\arguments{
\item{name}{dimension name}
}
\value{
a Druid dimension object
}
\description{
Creates a Druid dimension object
}

