\name{druid.build.aggregation}
\alias{druid.build.aggregation}
\title{Construct an arbitrary Druid aggregator using the given parameters}
\usage{
  druid.build.aggregation(type, ...)
}
\arguments{
  \item{type}{aggregator type}

  \item{...}{aggregator arguments}
}
\description{
  Construct an arbitrary Druid aggregator using the given
  parameters
}

