% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{druid.count}
\alias{druid.count}
\title{Constructs a count aggregator}
\usage{
druid.count(name = "count")
}
\arguments{
\item{name}{aggregator name, defaults to 'count'}
}
\description{
Constructs a count aggregator
}

