% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{druid.filter.and}
\alias{druid.filter.and}
\title{Construct an `and' filter of the given filters}
\usage{
druid.filter.and(...)
}
\arguments{
\item{...}{filters to combine}
}
\description{
Construct an `and' filter of the given filters
}

