\name{druid.filter.javascript}
\alias{druid.filter.javascript}
\title{Construct a JavaScript filter for a given dimension}
\usage{
  druid.filter.javascript(dimension, predicate)
}
\arguments{
  \item{dimension}{dimension to match}

  \item{predicate}{Javascript predicate function used to
  filter the dimension}
}
\description{
  Construct a JavaScript filter for a given dimension
}

