% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{druid.filter.not}
\alias{druid.filter.not}
\title{Construct a `not' filter of the given filter}
\usage{
druid.filter.not(filter)
}
\arguments{
\item{filter}{to negate}
}
\description{
Construct a `not' filter of the given filter
}

