% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{druid.filter.or}
\alias{druid.filter.or}
\title{Construct an `or' filter of the given filters}
\usage{
druid.filter.or(...)
}
\arguments{
\item{...}{filters to combine}
}
\description{
Construct an `or' filter of the given filters
}

