% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/having.R
\name{druid.having.not}
\alias{druid.having.not}
\title{Construct a `not' having of the given havingspec}
\usage{
druid.having.not(having)
}
\arguments{
\item{having}{to negate}
}
\description{
Construct a `not' having of the given havingspec
}

