% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postaggregator.R
\name{druid.postagg.constant}
\alias{druid.postagg.constant}
\title{Creates a constant post-aggregator}
\usage{
druid.postagg.constant(x)
}
\arguments{
\item{x}{numeric constant}
}
\description{
Creates a constant post-aggregator
}

