\name{druid.query.dataSources}
\alias{druid.query.dataSources}
\title{Query Druid data sources}
\usage{
  druid.query.dataSources(url = druid.url())
}
\arguments{
  \item{url}{URL to connect to druid, defaults to
  druid.url()}
}
\value{
  a character vector with the list of data sources
}
\description{
  Query Druid data sources
}

