% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{druid.query.dimensions}
\alias{druid.query.dimensions}
\title{Query data source dimensions}
\usage{
druid.query.dimensions(url = druid.url(), dataSource, interval = NULL, ...)
}
\arguments{
\item{url}{URL to connect to druid, defaults to druid.url()}

\item{dataSource}{name of the data source to query}
}
\value{
a character vector with the list of dimensions
}
\description{
Query data source dimensions
}

