% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{druid.query.select}
\alias{druid.query.select}
\title{Query raw Druid rows from a datasource}
\usage{
druid.query.select(url = druid.url(), dataSource, intervals,
  dimensions = list(), metrics = list(), descending = F, threshold,
  rawData = F, verbose = F, ...)
}
\arguments{
\item{url}{URL to connect to druid, defaults to druid.url()}

\item{dataSource}{name of the data source to query}

\item{intervals}{the time period to retrieve data for as an interval or list of interval objects}

\item{dimensions}{list of dimensions to include in query. An empty list returns all dimensions}

\item{metrics}{list of metrics to include in query. An empty list returns all metrics.}

\item{descending}{orders the data descending by date. Default is false.}

\item{threshold}{the maximum number of results the query can return}
}
\value{
Returns a dataframe with the raw Druid rows
}
\description{
For a given data source, get the raw Druid rows for a particular period
}

