% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{druid.selecttodf}
\alias{druid.selecttodf}
\title{Convert Druid select query result to a data frame}
\usage{
druid.selecttodf(result)
}
\arguments{
\item{result}{The result from query}
}
\description{
Retrieves the JSON result from Druid, then formats it into a dataframe with
a timestamp column
}
\seealso{
\code{\link{query}}
}

