% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{druid.url}
\alias{druid.url}
\title{Construct the URL for a druid endpoint.}
\usage{
druid.url(host = "localhost", port = 8082)
}
\arguments{
\item{host}{hostname}

\item{port}{port}
}
\value{
The URL for the druid endpoint
}
\description{
Defaults to http://localhost:8082/druid/v2/
}

