% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{==.druid.dimension}
\alias{==.druid.dimension}
\title{Construct a selector filter for a given dimension}
\usage{
\method{==}{druid.dimension}(a, b)
}
\arguments{
\item{a}{dimension to match}

\item{b}{value to match}
}
\description{
Construct a selector filter for a given dimension
}

