% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/having.R
\name{equalTo}
\alias{equalTo}
\title{Construct an equalTo having clause of the given named arguments}
\usage{
equalTo(...)
}
\arguments{
\item{...}{named arguments of equalTos to match}
}
\description{
Construct an equalTo having clause of the given named arguments
}

