% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postaggregator.R
\name{field}
\alias{field}
\title{Creates a fieldAccess post-aggregator}
\usage{
field(x)
}
\arguments{
\item{x}{name of the field}
}
\description{
Defines a reference to a Druid field (e.g. an aggregator)
}

