% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-time.R
\name{fromISO}
\alias{fromISO}
\title{Convert an ISO 8601 string to a POSIXct object}
\usage{
fromISO(x)
}
\arguments{
\item{x}{character string to be converted}
}
\description{
Convert an ISO 8601 string to a POSIXct object
}

