% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%=\%.druid.dimension}
\alias{\%=\%.druid.dimension}
\title{Construct a selector filter for a given dimension}
\usage{
\method{\%=\%}{druid.dimension}(dimension, x)
}
\arguments{
\item{dimension}{dimension to match}

\item{x}{character vector of possible values to match}
}
\description{
Construct a selector filter for a given dimension
}

