% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{hyperUnique}
\alias{hyperUnique}
\title{Constructs a hyperUnique aggregator over the given Druid metric}
\usage{
hyperUnique(x, ...)
}
\arguments{
\item{x}{metric}

\item{...}{optional arguments to construct aggregator}
}
\description{
Constructs a hyperUnique aggregator over the given Druid metric
}

