% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{longSum.druid.metric}
\alias{longSum.druid.metric}
\title{Constructs a longSum aggregator over the given Druid metric}
\usage{
\method{longSum}{druid.metric}(x, ...)
}
\arguments{
\item{x}{metric}

\item{...}{optional arguments to construct aggregator}
}
\description{
Constructs a longSum aggregator over the given Druid metric
}

