% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{max.druid.metric}
\alias{max.druid.metric}
\title{Constructs a max aggregator over the given Druid metric}
\usage{
\method{max}{druid.metric}(x, ...)
}
\arguments{
\item{x}{metric}

\item{...}{optional arguments to construct aggregator}
}
\description{
Constructs a max aggregator over the given Druid metric
}

