% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregator.R
\name{min.druid.metric}
\alias{min.druid.metric}
\title{Constructs a min aggregator over the given Druid metric}
\usage{
\method{min}{druid.metric}(x, ...)
}
\arguments{
\item{x}{metric}

\item{...}{optional arguments to construct aggregator}
}
\description{
Constructs a min aggregator over the given Druid metric
}

