% Generated by roxygen2 (4.0.2): do not edit by hand
\name{+.druid.postaggregator}
\alias{+.druid.postaggregator}
\title{Define a post-aggregator by adding aggregators, post-aggregators or constants}
\usage{
\method{+}{druid.postaggregator}(a, b)
}
\arguments{
\item{a}{aggregator, post-aggregator, or constant}

\item{b}{a aggregator, post-aggregator, or constant}
}
\value{
post-aggregator a + b
}
\description{
Define a post-aggregator by adding aggregators, post-aggregators or constants
}

