% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{renameagg}
\alias{renameagg}
\title{Rename aggregators based on their list name}
\usage{
renameagg(agglist)
}
\arguments{
\item{agglist}{list of aggregators}
}
\description{
Rename aggregators based on their list name
}

