% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postaggregator.R
\name{/.druid.postaggregator}
\alias{/.druid.postaggregator}
\title{Define a post-aggregator by dividing aggregators, post-aggregators or constants}
\usage{
\method{/}{druid.postaggregator}(a, b)
}
\arguments{
\item{a}{aggregator, post-aggregator, or constant}

\item{b}{a aggregator, post-aggregator, or constant}
}
\value{
post-aggregator a / b
}
\description{
Define a post-aggregator by dividing aggregators, post-aggregators or constants
}

