% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postaggregator.R
\name{sum.druid.postaggregator}
\alias{sum.druid.postaggregator}
\title{Sum of Druid post-aggregators}
\usage{
\method{sum}{druid.postaggregator}(...)
}
\arguments{
\item{...}{post-aggregators}
}
\description{
Sum of Druid post-aggregators
}

