% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toISO.Interval}
\alias{toISO.Interval}
\title{Convert a Interval object to its ISO 8601 string representation}
\usage{
\method{toISO}{Interval}(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a Interval object to its ISO 8601 string representation
}

