% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toISO.list}
\alias{toISO.list}
\title{Convert a list of objects to their ISO 8601 string representation}
\usage{
\method{toISO}{list}(x)
}
\arguments{
\item{x}{list to be converted}
}
\description{
Convert a list of objects to their ISO 8601 string representation
}

