% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-time.R
\name{toISO.numeric}
\alias{toISO.numeric}
\title{Convert a POSIX timestamp to its ISO 8601 string representation}
\usage{
\method{toISO}{numeric}(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a POSIX timestamp to its ISO 8601 string representation
}

