% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-time.R
\name{toISO.POSIXt}
\alias{toISO.POSIXt}
\title{Convert a POSIX* object to its ISO 8601 string representation}
\usage{
\method{toISO}{POSIXt}(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a POSIX* object to its ISO 8601 string representation
}

