% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postaggregator.R
\name{wrapConstant}
\alias{wrapConstant}
\title{helper function to wrap constants in formulas}
\usage{
wrapConstant(x)
}
\arguments{
\item{x}{numeric constant or post-aggregator}
}
\description{
helper function to wrap constants in formulas
}

