% Generated by roxygen2 (4.0.2): do not edit by hand
\name{druid.build.postaggregator}
\alias{druid.build.postaggregator}
\title{Creates a Druid post-aggregator}
\usage{
druid.build.postaggregator(type, ...)
}
\arguments{
\item{type}{post-aggregator type}

\item{...}{additional post-aggregator parameters}
}
\value{
a Druid post-aggregator object
}
\description{
Creates a Druid post-aggregator
}

