% Generated by roxygen2 (4.0.2): do not edit by hand
\name{druid.filter.regex}
\alias{druid.filter.regex}
\title{Construct a regex filter for a given dimension}
\usage{
druid.filter.regex(dimension, pattern)
}
\arguments{
\item{dimension}{dimension to match}

\item{pattern}{pattern to match}
}
\description{
Construct a regex filter for a given dimension
}

