% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/having.R
\name{druid.having.or}
\alias{druid.having.or}
\title{Construct an `or' having of the given havingspecs}
\usage{
druid.having.or(...)
}
\arguments{
\item{...}{havings to combine}
}
\description{
Construct an `or' having of the given havingspecs
}

