% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postaggregator.R
\name{druid.postagg.arithmetic}
\alias{druid.postagg.arithmetic}
\title{Creates an arithmetic post-aggregator}
\usage{
druid.postagg.arithmetic(fn, a, b, name = NULL)
}
\arguments{
\item{fn}{operator (e.g. +, -, *, /)}

\item{a}{left hand side post-aggregator}

\item{b}{right hand side post-aggregator}

\item{name}{alias for this post-aggregator}
}
\description{
Creates an arithmetic post-aggregator
}

