% Generated by roxygen2 (4.0.2): do not edit by hand
\name{druid.postagg.fieldAccess}
\alias{druid.postagg.fieldAccess}
\title{Creates a fieldAccess post-aggregator}
\usage{
druid.postagg.fieldAccess(name, fieldName = name)
}
\arguments{
\item{name}{alias for this post-aggregator}

\item{fieldName}{underlying field this post-aggregator refers to (defaults to name)}
}
\description{
Creates a fieldAccess post-aggregator
}

