\name{druid.query.dimensions}
\alias{druid.query.dimensions}
\title{Query data source dimensions}
\usage{
  druid.query.dimensions(url = druid.url(), dataSource)
}
\arguments{
  \item{url}{URL to connect to druid, defaults to
  druid.url()}

  \item{dataSource}{name of the data source to query}
}
\value{
  a character vector with the list of dimensions
}
\description{
  Query data source dimensions
}

