% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{\%=\%.druid.dimension}
\alias{\%=\%.druid.dimension}
\title{Construct a selector filter for a given dimension}
\usage{
\method{\%=\%}{druid.dimension}(dimension, pattern)
}
\arguments{
\item{dimension}{dimension to match}

\item{x}{character vector of possible values to match}
}
\description{
Construct a selector filter for a given dimension
}

