% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{\%~\%.druid.dimension}
\alias{\%~\%.druid.dimension}
\title{Construct a regex filter for a given dimension}
\usage{
\method{\%~\%}{druid.dimension}(dimension, pattern)
}
\arguments{
\item{dimension}{dimension to match}

\item{pattern}{pattern to match}
}
\description{
Construct a regex filter for a given dimension
}

