% Generated by roxygen2 (4.0.2): do not edit by hand
\name{>=.druid.aggregation}
\alias{>=.druid.aggregation}
\title{Construct a greaterThan having clause for a given aggregation}
\usage{
\method{>=}{druid.aggregation}(a, b)
}
\arguments{
\item{a}{aggregation to match}

\item{b}{value to match}
}
\description{
Construct a greaterThan having clause for a given aggregation
}

