\name{json}
\alias{json}
\title{calls toJSON with a default set of arguments}
\usage{
  json(obj, ...)
}
\arguments{
  \item{obj}{object to pass to toJSON}

  \item{...}{other arguments to pass to toJSON}
}
\description{
  - sets the default number of digits to 22 to retain
  maximum precisioin - removes newlines and spaces from
  json string
}

