% Generated by roxygen2 (4.0.2): do not edit by hand
\name{longSum}
\alias{longSum}
\title{Constructs a longSum aggregator over the given Druid metric}
\usage{
longSum(x, ...)
}
\arguments{
\item{x}{metric}

\item{...}{optional arguments to construct aggregator}
}
\description{
Constructs a longSum aggregator over the given Druid metric
}

