\name{min.druid.metric}
\alias{min.druid.metric}
\title{Constructs a min aggregator over the given Druid metric}
\usage{
  \method{min}{druid.metric} (x, ...)
}
\arguments{
  \item{x}{metric}

  \item{...}{optional arguments to construct aggregator}
}
\description{
  Constructs a min aggregator over the given Druid metric
}

