% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selector}
\alias{selector}
\title{Construct a selector filter of the given named arguments}
\usage{
selector(...)
}
\arguments{
\item{...}{named arguments of selectors to match}
}
\description{
Construct a selector filter of the given named arguments
}

