% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{selector}
\alias{selector}
\title{Construct a selector filter of the given named arguments}
\usage{
selector(...)
}
\arguments{
\item{...}{named arguments of selectors to match}
}
\description{
Construct a selector filter of the given named arguments
}

