\name{sum.druid.metric}
\alias{sum.druid.metric}
\title{Constructs a doubleSum aggregator over the given Druid metric}
\usage{
  \method{sum}{druid.metric} (x, ...)
}
\arguments{
  \item{x}{metric}

  \item{...}{optional arguments to construct aggregator}
}
\description{
  Constructs a doubleSum aggregator over the given Druid
  metric
}

