% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toISO.character}
\alias{toISO.character}
\title{toISO leaves strings as is without checking for ISO 8601 validity}
\usage{
\method{toISO}{character}(x)
}
\arguments{
\item{x}{ISO 8601 character vector}
}
\description{
toISO leaves strings as is without checking for ISO 8601 validity
}

