% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toISO.numeric}
\alias{toISO.numeric}
\title{Convert a POSIX timestamp to its ISO 8601 string representation}
\usage{
\method{toISO}{numeric}(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a POSIX timestamp to its ISO 8601 string representation
}

