% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toISO.POSIXt}
\alias{toISO.POSIXt}
\title{Convert a POSIX* object to its ISO 8601 string representation}
\usage{
\method{toISO}{POSIXt}(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a POSIX* object to its ISO 8601 string representation
}

